Profile View Refactor. Profile View Model Creation
