--- DubDubGrub/Views/Screens/ProfileView.swift
@@ -1,229 +1,0 @@
-//
-//  ProfileView.swift
-//  DubDubGrub
-//
-//  Created by Carlos Reyes on 6/21/25.
-//
-
-import SwiftUI
-import CloudKit
-
-struct ProfileView: View {
-    
-    @State private var firstName    = ""
-    @State private var lastName     = ""
-    @State private var companyName  = ""
-    @State private var bio          = ""
-    @State private var avatar       = PlaceholderImage.avatar
-    @State private var isShowingPhotoPicker = false
-    @State private var alertItem: AlertItem?
-    @State private var alertIsPresented: Bool = false
-    
-    var body: some View {
-        VStack {
-            ZStack {
-                NameBackgroundView()
-                
-                HStack(spacing: 16) {
-                    ZStack {
-                        AvatarView(image: avatar, size: 84)
-                        EditImage()
-                    }
-                    .padding(.leading, 12)
-                    .onTapGesture { isShowingPhotoPicker = true }
-                    
-                    VStack(spacing:1) {
-                        TextField("First Name", text: $firstName)
-                            .profileNameStyle()
-                        
-                        TextField("Last Name", text: $lastName)
-                            .profileNameStyle()
-                        
-                        TextField("Company Name", text: $companyName)
-                    }
-                    .padding(.trailing, 16)
-                }
-                .padding()
-            }
-            
-            VStack(alignment: .leading, spacing: 8) {
-                CharactersRemainView(currentCount: bio.count)
-                
-                TextEditor(text: $bio)
-                    .frame(height: 100)
-                    .overlay {
-                        RoundedRectangle(cornerRadius: 8).stroke(.secondary, lineWidth: 1)
-                    }
-            }
-            .padding(.horizontal, 16)
-            
-            Spacer()
-            
-            Button {
-                createProfile()
-            } label: {
-                DDGButton(title: "Create Profile")
-            }
-            .padding(.bottom)
-        }
-        .navigationTitle("Profile")
-        .toolbar {
-            Button {
-                dismissKeyboard()
-            } label: {
-                Image(systemName: "keyboard.chevron.compact.down")
-            }
-        }
-        .onAppear {
-            getProfile()
-        }
-        .alert(alertItem?.title ?? "no alert",
-               isPresented: $alertIsPresented,
-               actions: { } ,
-               message: { alertItem?.message ?? Text("no message") }
-        )
-        .sheet(isPresented: $isShowingPhotoPicker) {
-            PhotoPicker(image: $avatar)
-        }
-    }
-    
-    func isValidProfile() -> Bool {
-        guard !firstName.isEmpty,
-              !lastName.isEmpty,
-              !companyName.isEmpty,
-              !bio.isEmpty,
-              avatar != PlaceholderImage.avatar,
-              bio.count < 100 else { return false }
-        
-        return true
-    }
-    
-    func createProfile() {
-        guard isValidProfile() else {
-            alertItem = AlertContext.invalidProfile
-            alertIsPresented = true
-            return
-        }
-        
-        // Create our CKRecord from the profile view
-        let profileRecord = CKRecord(recordType: RecordType.profile)
-        profileRecord[DDGProfile.kFirstName] = firstName
-        profileRecord[DDGProfile.kLastName] = lastName
-        profileRecord[DDGProfile.kCompanyName] = companyName
-        profileRecord[DDGProfile.kBio] = bio
-        profileRecord[DDGProfile.kAvatar] = avatar.convertToCKAsset()
-        
-        // Get our UserRecordId from the Container
-        CKContainer.default().fetchUserRecordID { recordID, error in
-            guard let recordID = recordID, error == nil else {
-                print(error!.localizedDescription)
-                return
-            }
-            
-            // Get UserRecord from the Public Database
-            CKContainer.default().publicCloudDatabase.fetch(withRecordID: recordID) { userRecord, error in
-                guard let userRecord = userRecord, error == nil else {
-                    print(error!.localizedDescription)
-                    return
-                }
-                
-                // Create reference on UserRecord to the DDGProfile we created
-                userRecord["userProfile"] = CKRecord.Reference(recordID: profileRecord.recordID, action: .none)
-                
-                // Create a CKOperation to save our User and Profile Records
-                let operation = CKModifyRecordsOperation(recordsToSave: [userRecord, profileRecord])
-                
-                operation.modifyRecordsResultBlock = { result in
-                    switch result {
-                    case .success():
-                        print("Succesfully created and uploaded profile to CloudKit")
-                    case .failure(let error):
-                        print(error.localizedDescription)
-                    }
-                }
-                
-                CKContainer.default().publicCloudDatabase.add(operation)
-            }
-        }
-    }
-    
-    func getProfile() {
-        CKContainer.default().fetchUserRecordID { recordID, error in
-            guard let recordID = recordID, error == nil else {
-                print(error!.localizedDescription)
-                return
-            }
-            
-            CKContainer.default().publicCloudDatabase.fetch(withRecordID: recordID) { userRecord, error in
-                guard let userRecord = userRecord, error == nil else {
-                    print(error!.localizedDescription)
-                    return
-                }
-                
-                let profileReference = userRecord["userProfile"] as! CKRecord.Reference
-                let profileRecordID = profileReference.recordID
-                
-                CKContainer.default().publicCloudDatabase.fetch(withRecordID: profileRecordID) { profileRecord, error in
-                    guard let profileRecord = profileRecord, error == nil else {
-                        print(error!.localizedDescription)
-                        return
-                    }
-                    
-                    DispatchQueue.main.async {
-                        let profile = DDGProfile(record: profileRecord)
-                        firstName   = profile.firstName
-                        lastName    = profile.lastName
-                        companyName = profile.companyName
-                        bio         = profile.bio
-                        avatar      = profile.createAvatarImage()
-                    }
-                }
-            }
-        }
-    }
-}
-
-#Preview {
-    NavigationView {
-        ProfileView()
-    }
-}
-
-struct NameBackgroundView: View {
-    var body: some View {
-        Color(.secondarySystemBackground)
-            .frame(height: 130)
-            .cornerRadius(12)
-            .padding()
-    }
-}
-
-struct EditImage: View {
-    var body: some View {
-        Image(systemName: "square.and.pencil")
-            .resizable()
-            .scaledToFit()
-            .frame(width: 14, height: 14)
-            .foregroundStyle(.white)
-            .offset(y: 30)
-    }
-}
-
-struct CharactersRemainView: View {
-    var currentCount: Int
-    
-    var body: some View {
-        Text("Bio: ")
-            .font(.callout)
-            .foregroundStyle(.secondary)
-        +
-        Text("\(100 - currentCount)")
-            .bold()
-            .font(.callout)
-            .foregroundStyle(currentCount <= 100 ? .brandPrimary : .pink)
-        +
-        Text(" Characters remain")
-            .font(.callout)
-            .foregroundStyle(.secondary)
-    }
-}
