+++ DubDubGrub/Views/Screens/ProfileView/ProfileView.swift
@@ -1,0 +1,123 @@
+//
+//  ProfileView.swift
+//  DubDubGrub
+//
+//  Created by Carlos Reyes on 6/21/25.
+//
+
+import SwiftUI
+import CloudKit
+
+struct ProfileView: View {
+    
+    @StateObject private var viewModel = ProfileViewModel()
+    
+    var body: some View {
+        VStack {
+            ZStack {
+                NameBackgroundView()
+                
+                HStack(spacing: 16) {
+                    ZStack {
+                        AvatarView(image: viewModel.avatar, size: 84)
+                        EditImage()
+                    }
+                    .padding(.leading, 12)
+                    .onTapGesture { viewModel.isShowingPhotoPicker = true }
+                    
+                    VStack(spacing:1) {
+                        TextField("First Name", text: $viewModel.firstName).profileNameStyle()
+                        TextField("Last Name", text: $viewModel.lastName).profileNameStyle()
+                        TextField("Company Name", text: $viewModel.companyName)
+                    }
+                    .padding(.trailing, 16)
+                }
+                .padding()
+            }
+            
+            VStack(alignment: .leading, spacing: 8) {
+                CharactersRemainView(currentCount: viewModel.bio.count)
+                
+                TextEditor(text: $viewModel.bio)
+                    .frame(height: 100)
+                    .overlay {
+                        RoundedRectangle(cornerRadius: 8).stroke(.secondary, lineWidth: 1)
+                    }
+            }
+            .padding(.horizontal, 16)
+            
+            Spacer()
+            
+            Button {
+                viewModel.createProfile()
+            } label: {
+                DDGButton(title: "Create Profile")
+            }
+            .padding(.bottom)
+        }
+        .navigationTitle("Profile")
+        .toolbar {
+            Button {
+                dismissKeyboard()
+            } label: {
+                Image(systemName: "keyboard.chevron.compact.down")
+            }
+        }
+        .onAppear {
+            viewModel.getProfile()
+        }
+        .alert(viewModel.alertItem?.title ?? "no alert",
+               isPresented: $viewModel.alertIsPresented,
+               actions: { } ,
+               message: { viewModel.alertItem?.message ?? Text("no message") }
+        )
+        .sheet(isPresented: $viewModel.isShowingPhotoPicker) {
+            PhotoPicker(image: $viewModel.avatar)
+        }
+    }
+}
+
+#Preview {
+    NavigationView {
+        ProfileView()
+    }
+}
+
+struct NameBackgroundView: View {
+    var body: some View {
+        Color(.secondarySystemBackground)
+            .frame(height: 130)
+            .cornerRadius(12)
+            .padding()
+    }
+}
+
+struct EditImage: View {
+    var body: some View {
+        Image(systemName: "square.and.pencil")
+            .resizable()
+            .scaledToFit()
+            .frame(width: 14, height: 14)
+            .foregroundStyle(.white)
+            .offset(y: 30)
+    }
+}
+
+struct CharactersRemainView: View {
+    var currentCount: Int
+    
+    var body: some View {
+        Text("Bio: ")
+            .font(.callout)
+            .foregroundStyle(.secondary)
+        +
+        Text("\(100 - currentCount)")
+            .bold()
+            .font(.callout)
+            .foregroundStyle(currentCount <= 100 ? .brandPrimary : .pink)
+        +
+        Text(" Characters remain")
+            .font(.callout)
+            .foregroundStyle(.secondary)
+    }
+}
