+++ DubDubGrub/Views/Screens/ProfileView/ProfileViewModel.swift
@@ -1,0 +1,122 @@
+//
+//  ProfileViewModel.swift
+//  DubDubGrub
+//
+//  Created by Carlos Reyes on 10/27/25.
+//
+
+import CloudKit
+
+final class ProfileViewModel: ObservableObject {
+    
+    @Published var firstName    = ""
+    @Published var lastName     = ""
+    @Published var companyName  = ""
+    @Published var bio          = ""
+    @Published var avatar       = PlaceholderImage.avatar
+    @Published var isShowingPhotoPicker = false
+    @Published var alertItem: AlertItem?
+    @Published var alertIsPresented: Bool = false
+    
+    func isValidProfile() -> Bool {
+        guard !firstName.isEmpty,
+              !lastName.isEmpty,
+              !companyName.isEmpty,
+              !bio.isEmpty,
+              avatar != PlaceholderImage.avatar,
+              bio.count < 100 else { return false }
+        
+        return true
+    }
+    
+    func createProfile() {
+        guard isValidProfile() else {
+            alertItem = AlertContext.invalidProfile
+            alertIsPresented = true
+            return
+        }
+
+        
+        // Get our UserRecordId from the Container
+        CKContainer.default().fetchUserRecordID { recordID, error in
+            guard let recordID = recordID, error == nil else {
+                print(error!.localizedDescription)
+                return
+            }
+            
+            let profileRecord = self.createProfileRecord()
+            
+            // Get UserRecord from the Public Database
+            CKContainer.default().publicCloudDatabase.fetch(withRecordID: recordID) { userRecord, error in
+                guard let userRecord = userRecord, error == nil else {
+                    print(error!.localizedDescription)
+                    return
+                }
+                
+                // Create reference on UserRecord to the DDGProfile we created
+                userRecord["userProfile"] = CKRecord.Reference(recordID: profileRecord.recordID, action: .none)
+                
+                // Create a CKOperation to save our User and Profile Records
+                let operation = CKModifyRecordsOperation(recordsToSave: [userRecord, profileRecord])
+                
+                operation.modifyRecordsResultBlock = { result in
+                    switch result {
+                    case .success():
+                        print("Succesfully created and uploaded profile to CloudKit")
+                    case .failure(let error):
+                        print(error.localizedDescription)
+                    }
+                }
+                
+                CKContainer.default().publicCloudDatabase.add(operation)
+            }
+        }
+    }
+    
+    func getProfile() {
+        CKContainer.default().fetchUserRecordID { recordID, error in
+            guard let recordID = recordID, error == nil else {
+                print(error!.localizedDescription)
+                return
+            }
+            
+            CKContainer.default().publicCloudDatabase.fetch(withRecordID: recordID) { userRecord, error in
+                guard let userRecord = userRecord, error == nil else {
+                    print(error!.localizedDescription)
+                    return
+                }
+                
+                let profileReference = userRecord["userProfile"] as! CKRecord.Reference
+                let profileRecordID = profileReference.recordID
+                
+                CKContainer.default().publicCloudDatabase.fetch(withRecordID: profileRecordID) { profileRecord, error in
+                    guard let profileRecord = profileRecord, error == nil else {
+                        print(error!.localizedDescription)
+                        return
+                    }
+                    
+                    DispatchQueue.main.async {
+                        let profile = DDGProfile(record: profileRecord)
+                        self.firstName   = profile.firstName
+                        self.lastName    = profile.lastName
+                        self.companyName = profile.companyName
+                        self.bio         = profile.bio
+                        self.avatar      = profile.createAvatarImage()
+                    }
+                }
+            }
+        }
+    }
+    
+    private func createProfileRecord() -> CKRecord {
+        
+        let profileRecord = CKRecord(recordType: RecordType.profile)
+        profileRecord[DDGProfile.kFirstName]    = firstName
+        profileRecord[DDGProfile.kLastName]     = lastName
+        profileRecord[DDGProfile.kCompanyName]  = companyName
+        profileRecord[DDGProfile.kBio]          = bio
+        profileRecord[DDGProfile.kAvatar]       = avatar.convertToCKAsset()
+        
+        return profileRecord
+    }
+}
